Remove static link in Android framework under /frameworks/native.
The dependencies to static libraries in frameworks should be removed
as many as possible to reduce a size of the system partition.
And that will improve coverage of the VNDK libraries because this
effort might find new necessary shared libs which were linked
statically before.

Size diffs of affected binaries on angler-userdebug build are as
follows.

dumpstate            : 371048 -> 325312 (-45736)
libvulkan.so         : 104892 -> 92016 (-12876)
libandroid_runtime.so: 1193240 -> 1193296 (+56)

total                : (-58556)

Test: building succeeded, and the image was tested on angler.
Bug: 33056637
Change-Id: I2db1e94d8bc9cad64760cdf36cac445273bd641a