Fix installing dumpstate test data
The dumpstate test data was being copied to
$OUT/data/nativetest/dumpstate_test_fixture, but then
LOCAL_PICKUP_FILES was set to $OUT/data, which pulls in everything
that has been installed to /data.  This results in lots of extra
files in the *tests.zip files.

Copy the test data to the intermediates directory for use with
LOCAL_PICKUP_FILES, and then again again to $OUT/data for use with
adb sync.

Bug: 30564705
Bug: 31807540
Bug: 31997691
Test: mma -j, test data is installed to $OUT/data and intermediates/DATA
Test: adb sync && adb shell /data/nativetest/dumpstate_test/dumpstate_test
Change-Id: Iaa83d96e5d23788b301fed8819a446baf127b58e