Revert "atrace: Poke PDX services using utility class."
This reverts commit f70680e995d6df4604c9b34f6a5afebe613d7feb.

There appears to be a bug in the current build system that does not
cleanly rebuild all targets using static libraries built across
different build systems (Soong/make). With the recent change to PDX
connection handshake, it is important that all clients and services are
rebuilt to include the new protocol changes. However this is not the
case when doing incremental builds. Some services remain untouched
during global builds and hence there is a disparity in clients' and
services' implementations of the protocol.

atrace being one of the clients is affected (the symptom is that when
started atrace just hangs waiting for connection to a PDX service).

While there is nothing wrong with the current state of code, but the
issue of incomplete incremental builds is causing headaches for people
trying to chase performance issues in O with systrace (since it is
unclear that they need to do a clean build).

Reverting the dependency on PDX in atrace for now until the build system
issue is identified and fixed property. This change can be re-landed then.

Bug: 38201226
Change-Id: Ica0efc38f600295ef6f64b9f9584ac41cb84f424