Use logcat -b kernel for bug reports
One of the main advantages of the kernel log buffer is that _spammy_
selinux violations are suppressed there, and instead land in events and
main log buffers for all to view. This gives the kernel log buffer a
longer logspan and better signal to noise ratio than an equivalently
sized dmesg buffer. The other advantage is the CLOCK_REALTIME usage,
that makes it easier to correlate kernel and user
space activities in the bugreport.

Bug: 30736473
Test: 1) adb bugreport 2) adb shell setprop ro.logd.kernel false
3) adb bugreport
Then compare the two resulting bug reports (open both in Android
Bug Tool). Observe that timestamps are seconds since bootup
in the second bugreport, and actual time (human-readable) in the first
bugreport.

Change-Id: I968de323833dda97ded4ecc454e12220d4bd3021