Build libbufferhubqueue and libpdx_default_transport as shared library
Now that we are moving BufferHub into libgui, these two libraries are
being linked into libgui.so and libdvr.so respectively as static
lib. This is cauing issues when a binary/app depends on both, i.e. two
copies of the static lib will live in the same binary. We can avoid this
issue by building these two as shared library.

Test: Build system
Bug: 72763929
Change-Id: I26ba18f2b84d43cbd62b26dbb9ee0574c2974f7d