lshal: mention subclasses for third section
Third section of lshal shows installed passthrough
services. I've updated the comment to show that
these aren't necessarily the actual classes available
at runtime. This is based on an offline issue raised
by a vendor that thought the current text was confusing.

The actual check that the correct passthrough services
are served is done by vts_treble_vintf_test.

Bug: 74407622
Test: lshal
Change-Id: Ie314c7a4bd9abe9d52ca36cad1bf8da80493b3eb