Remove ATRACE_TAG_APP from atrace's control
We control app tracing via the -a option, which system properties that
early app startup code inspects. This startup code ATRACE_TAG_APP
internally depending on the value of the properties derived from the
-a space, so letting users control the global value of ATRACE_TAG_APP
from atrace provides a false sense of control.

This CL removes the ability to set the global tag, directing user
attention to the -a option instead.

Test: code inspection
Change-Id: If981ba302db27da1c9f1942cf54d400bd4672b73