Remove ftrace dumping from dumpstate
When capturing a bug report, dumpstate checks if ftrace
is on, and if so, steals the trace from the buffer.
This feature doesn't seem to be actually used, getting
rid of unnecessary complexity and unnecessary permissions.
Furthermore, stealing the trace is bad for iorapd, studio
perfetto, Traceur and all the other trace users.
Matching selinux change: aosp/875395

Bug: 122987614
Test: run dumpstate before and after patch,
      file sizes are comparable,
      observed no tracing-related errors.
Change-Id: I8bcd478c9bde776a2ec21b5e9a97aedaaa3681fd