Move basic events into the sched category in atrace.
Several of these basic events were in the 'memory' category because we
were initially finding them most useful for analyzing memory. However,
they're actually useful across the board, and should be included in
every trace. They're more appropriate in the 'sched' category.

Also, noticed that sched_process_exit wasn't writeable because it needed
to be added to the atrace.rc file, so fixed that.

Bug: 130416576
Test: manual: ran atrace with only the 'sched' category and saw these
tracepoints recorded.

Change-Id: Ib6bd958c1b873acfa8b04a79fc2b9b65a529d768