dumpstate: Only attempt screencap with do_fb
When 'do_fb' is false, then our 'screenshot_path_' is empty,
so our call to TakeScreenshot() results in two Errors into
the logcat.  We want to avoid that, especially during the
bugreport process.

So we follow the logic which 'do_early_screenshot_' uses, by
only calling TakeScreenshot() when 'do_fb' is true.  We also
modify this other call to TakeScreenshot() to remove its
additional check, since we know if we get into this unexpected
state, we'll get non-fatal Errors in the logcat indicating it.

Bug: 133448994
Test: dumpstate_test
Change-Id: Ibb30bb908e12a498a341e2c5c1d9b59a7a16c2a5