Prefactoring: rewrite section detection smoke tests
Smoke tests currently rely on onSectionComplete callbacks to test
bugreports generate sections. onSectionComplete methods are otherwise
unused and can be removed as such. In preparation to that migrate the
tests to parse the bugreport and identify sections.

While at it, also fix ContainsMainEntry test, which was flaky because
the string read from main_entry.txt was not null terminated.

BUG: 128980174
Test: adb shell /data/nativetest64/dumpstate_smoke_test/dumpstate_smoke_test

Change-Id: Icf87f80378d5eced7d0ceafeef6a30bb9ecd0871