Stop attempting to read /d/sync
The info in /d/sync was relocated to a different file in linux kernel
version 4.8 so this read has been returning error for a while now.
Also, all debugfs usage from platform code is being cleaned up in
Android R due to the instability of its ABI and dumpstate is no longer
permitted to access debugfs.

Bug: 134669095
Test: make
Change-Id: I4fc559229b1ac5dea7371dfa7e83af173b5feb9b