Use lower filesystem path when destroying user data with FUSE
With the FUSE prop enabled, when a user is removed, the upper
filesystem path would have been unmounted before installd gets to
delete its content and an empty tmpfs will be deleted instead, leaving
the content on the lower filesystem.

Now, we always use the lower filesystem when deleting the users dir
and cache. This is fine on devices without sdcardfs but on devices
running sdcardfs, we should use the sdcardfs path instead to avoid
VFS inode cache inconsistencies

Test: adb shell pm create-user && adb shell am start-user 10 && adb
shell pm delete-user. Verify that /data/media/10 is deleted
Bug: 135341433
Change-Id: Iaeb9ea3db373d358251644a95b3390cfa2b8f24a