[cmd_fuzzer] Depend on libutils.so instead of libutils.a
Bug: http://b/197965342

libbinder.so, a dependency of cmd_fuzzer, also depends on libutils.so.
If we include libutils.a when linking cmd_fuzzer, upstream's lld linker
will unnecessarily fetch symbols from libutils.a to redefine undefined
symbols from libbinder.so.  Switching to libutils.so does not increase
the closure of shared libraries needed for this test.

Test: build with aosp/1809741 which has lld with the above behavior.
Change-Id: Id11e1b95a968429b1406c65f51d2c70e197ae241