Stop creating sdk data directories due to boot time regression
For some reason, there is 180% increase in boot time due to creating of
sdk data directories. Locally, I noticed that it takes ~2500ms to prepare
data. If we stop preparing sdk data directories with this cl, that time
drops to ~1200ms.

Will revert this CL once we fix the root cause of the regression.

Bug: 220095381
Test: manual
Ignore-AOSP-First: Feature is being developed in internal branch
Change-Id: I3b74e34344ceb161d584c98bcb816d3dff3fd392